Test del servicio de plato completado

+++ ReservaBook/ReservaBook.Core.Aplication/Dtos/plato/CreatePlatoRequestDto.cs
@@ -11,7 +11,7 @@ namespace ReservaBook.Core.Aplication.Dtos.plato
         public required string Nombre { get; set; }
         public required string Descripcion { get; set; }
         public string? Imagen { get; set; }
-        public required double Precio { get; set; }
+        public required decimal Precio { get; set; }
         public required string Categoria { get; set; }
         public required string Estado { get; set; }
