+++ ReservaBook/ReservaBook.Core.Domain/Entities/Plato.cs
@@ -12,7 +12,7 @@ namespace ReservaBook.Core.Domain.Entities
         public required string Nombre { get; set; }
         public required string Descripcion { get; set; }
         public string? Imagen {  get; set; }
-        public required double Precio { get; set; }
+        public required decimal Precio { get; set; }
         public required string Categoria { get; set; }
         public required string Estado { get; set; }
